HTCONDOR-2021: Make Python Utility Library
-Made new _utils submodule for htcondor (in python)
-Made ansii shared utilities for coloring purposes
    - Added to python utils lib
    - Updated condor_watch_q
    - Updated 'htcondor dag status'
-Fixed bug in condor_watch_q where -no-color flag didn't
 fall all the way through the code. Resulting in per row
 progress bars still being colored
